requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
yfinance>=1.2.0
pandas>=2.3.0
APIConnect==2.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
youtube-transcript-api>=0.6.2
